refactor: ai module core
