+++ apps/desktop/layer/renderer/package.json
@@ -13,8 +13,6 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@ai-sdk/openai": "2.0.0-beta.11",
-    "@ai-sdk/react": "2.0.0-beta.25",
     "@dnd-kit/core": "6.3.1",
     "@dnd-kit/sortable": "10.0.0",
     "@electron-toolkit/preload": "3.0.2",
@@ -53,7 +51,7 @@
     "@use-gesture/react": "10.3.1",
     "@welldone-software/why-did-you-render": "10.0.1",
     "@yornaath/batshit": "0.10.1",
-    "ai": "5.0.0-beta.25",
+    "ai": "5.0.0-beta.34",
     "camelcase-keys": "9.1.3",
     "class-variance-authority": "0.7.1",
     "clsx": "2.1.1",
