+++ apps/desktop/layer/renderer/src/modules/ai/chat/__internal__/slices/index.ts
@@ -1,5 +1,7 @@
+export type { ChatSlice } from "../chat-core/types"
 export {
   type BlockSlice as ContextSlice,
   createBlockSlice as createContextSlice,
 } from "./block.slice"
-export { type ChatSlice, type ChatStatus, createChatSlice } from "./chat.slice"
+export { createChatSlice } from "./chat.slice"
+export { type ChatStatus } from "ai"
