+++ apps/desktop/layer/renderer/src/modules/ai/chat/__internal__/store.ts
@@ -1,8 +1,8 @@
 import { createWithEqualityFn } from "zustand/traditional"
 
+import type { ChatSlice } from "./chat-core/types"
 import type { BlockSlice } from "./slices/block.slice"
 import { createBlockSlice } from "./slices/block.slice"
-import type { ChatSlice } from "./slices/chat.slice"
 import { createChatSlice } from "./slices/chat.slice"
 import type { AIChatStoreInitial } from "./types"
