+++ apps/desktop/layer/renderer/src/modules/ai/chat/components/layouts/ChatMoreDropdown.tsx
@@ -38,6 +38,7 @@ export const ChatMoreDropdown = () => {
   const handleDropdownOpen = (open: boolean) => {
     if (open) {
       loadHistory()
+      AIPersistService.cleanupEmptySessions()
     }
   }
